new smart account generation process refactored
